requirements.txt
python-dotenv==1.2.1
openai==2.15.0
python-telegram-bot==20.7
anyio==4.12.1
certifi==2026.1.4
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
jiter==0.12.0
pydantic==2.12.5
pydantic-core==2.41.5
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
